optimizer: Add suppport for UBOs and SSBOs
Ensure that our context has GLES 3.1 or GL_ARB_shader_storage_buffer_object
enabled, and initialize the relevant consts to sensible values (taken from the
mesa standalone compiler) so that input shaders with these blocks are parsed
correctly.

Insert an ir_typedecl_statement (previously only used for structs) in to the IR
for interface blocks. Ensure that ir_decl_removal_visitor only removes unused
structs rather than interface blocks too.

If the GL_ARB_shader_storage_buffer_object extension was enabled in the
input shader, ensure it is enabled in the output shader.

When printing the GLSL for an interface typedecl, obtain required
information (such as the variable mode and the binding) by searching the IR for
a variable that was either declared as the block or within the block. Ensure
that a variable that was declared as the block is also declared in the output
GLSL, and that unsized arrays are printed correctly.

Since variables declared as or within the interface block will be declared by
the typedecl, ensure that the output GLSL for the ir_variable is skipped in
these cases.